+++ plugins/source/aws/resources/services/servicecatalog/portfolios_mock_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go-v2/service/servicecatalog"
-	"github.com/aws/aws-sdk-go-v2/service/servicecatalogappregistry"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client/mocks"
 	"github.com/cloudquery/plugin-sdk/v4/faker"
@@ -14,7 +13,6 @@ import (
 
 func buildPortfolios(t *testing.T, ctrl *gomock.Controller) client.Services {
 	mk := mocks.NewMockServicecatalogClient(ctrl)
-	ma := mocks.NewMockServicecatalogappregistryClient(ctrl)
 
 	o := servicecatalog.ListPortfoliosOutput{}
 	require.NoError(t, faker.FakeObject(&o))
@@ -26,17 +24,16 @@ func buildPortfolios(t *testing.T, ctrl *gomock.Controller) client.Services {
 		nil,
 	)
 
-	to := servicecatalogappregistry.ListTagsForResourceOutput{}
-	require.NoError(t, faker.FakeObject(&to))
+	po := servicecatalog.DescribePortfolioOutput{}
+	require.NoError(t, faker.FakeObject(&po))
 
-	ma.EXPECT().ListTagsForResource(gomock.Any(), gomock.Any(), gomock.Any()).Return(
-		&to,
+	mk.EXPECT().DescribePortfolio(gomock.Any(), gomock.Any(), gomock.Any()).Return(
+		&po,
 		nil,
 	)
 
 	return client.Services{
-		Servicecatalog:            mk,
-		Servicecatalogappregistry: ma,
+		Servicecatalog: mk,
 	}
 }