+++ plugins/source/aws/resources/services/servicecatalog/provisioned_products.go
@@ -23,6 +23,7 @@ func ProvisionedProducts() *schema.Table {
 		Multiplex:   client.ServiceAccountRegionMultiplexer(tableName, "servicecatalog"),
 		Columns: []schema.Column{
 			client.DefaultAccountIDColumn(false),
+			client.DefaultRegionColumn(false),
 			{
 				Name:       "arn",
 				Type:       arrow.BinaryTypes.String,
@@ -32,8 +33,12 @@ func ProvisionedProducts() *schema.Table {
 			{
 				Name:     "tags",
 				Type:     sdkTypes.ExtensionTypes.JSON,
-				Resolver: resolveProvisionedProductTags,
+				Resolver: client.ResolveTags,
 			},
+		},
+		Relations: schema.Tables{
+			provisioningArtifact(),
+			launchPaths(),
 		},
 	}
 }
@@ -55,9 +60,4 @@ func fetchServicecatalogProvisionedProducts(ctx context.Context, meta schema.Cli
 	}
 
 	return nil
-}
-
-func resolveProvisionedProductTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
-	p := resource.Item.(types.ProvisionedProductAttribute)
-	return resource.Set(c.Name, client.TagsToMap(p.Tags))
 }