+++ plugins/source/k8s/client/testing.go
@@ -6,8 +6,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cloudquery/plugin-sdk/v4/plugin"
 	"github.com/cloudquery/plugin-sdk/v4/scheduler"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
+	"github.com/cloudquery/plugin-sdk/v4/transformers"
 	"github.com/golang/mock/gomock"
 	"github.com/rs/zerolog"
 	v1 "k8s.io/api/core/v1"
@@ -42,15 +44,15 @@ func K8sMockTestHelper(t *testing.T, table *schema.Table, builder func(*testing.
 		opt(c)
 	}
 	sched := scheduler.NewScheduler(scheduler.WithLogger(l))
-	messages, err := sched.SyncAll(context.Background(), c, schema.Tables{table})
+	tables := schema.Tables{table}
+	if err := transformers.TransformTables(tables); err != nil {
+		t.Fatal(err)
+	}
+	messages, err := sched.SyncAll(context.Background(), c, tables)
 	if err != nil {
 		t.Fatalf("failed to sync: %v", err)
 	}
-	records := messages.GetInserts().GetRecordsForTable(table)
-	emptyColumns := schema.FindEmptyColumns(table, records)
-	if len(emptyColumns) > 0 {
-		t.Fatalf("empty columns: %v", emptyColumns)
-	}
+	plugin.ValidateNoEmptyColumns(t, tables, messages)
 }
 
 func APIExtensionsMockTestHelper(t *testing.T, table *schema.Table, builder func(*testing.T, *gomock.Controller) apiextensionsclientset.Interface, opts ...TestOption) {