+++ plugins/source/mysql/resources/plugin/plugin_test.go
@@ -3,6 +3,7 @@ package plugin
 import (
 	"context"
 	"database/sql"
+	"encoding/json"
 	"fmt"
 	"os"
 	"sort"
@@ -12,9 +13,10 @@ import (
 
 	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/cloudquery/cloudquery/plugins/source/mysql/client"
-	"github.com/cloudquery/plugin-pb-go/specs"
-	"github.com/cloudquery/plugin-sdk/v3/schema"
-	"github.com/cloudquery/plugin-sdk/v3/types"
+	"github.com/cloudquery/plugin-sdk/v4/message"
+	"github.com/cloudquery/plugin-sdk/v4/plugin"
+	"github.com/cloudquery/plugin-sdk/v4/schema"
+	"github.com/cloudquery/plugin-sdk/v4/types"
 	"github.com/go-sql-driver/mysql"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/require"
@@ -111,60 +113,11 @@ func insertTable(ctx context.Context, db *sql.DB, table *schema.Table, records [
 	return nil
 }
 
-func getActualStringsForResource(resource *schema.Resource) []string {
-	actualData := resource.GetValues()
-	actualStrings := make([]string, len(actualData))
-	for i, v := range actualData {
-		// TODO: Remove this when https://bugs.mysql.com/bug.php?id=98135 is fixed
-		// We can't do string comparison for JSON columns as MySQL saves them with extra spaces
-		if v.DataType() != types.ExtensionTypes.JSON {
-			actualStrings[i] = v.String()
-		}
-	}
-	return actualStrings
-}
-
-func getActualStrings(resources []*schema.Resource) [][]string {
-	actualStrings := make([][]string, len(resources))
-	for i, resource := range resources {
-		actualStrings[i] = getActualStringsForResource(resource)
-	}
-	return actualStrings
-}
-
-func getExpectedStringsForRecord(t *testing.T, table *schema.Table, record arrow.Record) []string {
-	cloned := table.Copy(nil)
-	for i := range cloned.Columns {
-		// Normalize column types to the ones supported by MySQL
-		sqlType := client.SQLType(cloned.Columns[i].Type)
-		cloned.Columns[i].Type = client.SchemaType(sqlType, sqlType)
-	}
-
-	transformedRecord, _ := client.TransformRecord(record)
-	values := transformedRecord[0]
-	resource := schema.NewResourceData(cloned, nil, values)
-	for i, col := range cloned.Columns {
-		if err := resource.Set(col.Name, values[i]); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	return getActualStringsForResource(resource)
-}
-
-func getExpectedStrings(t *testing.T, table *schema.Table, records []arrow.Record) [][]string {
-	expectedStrings := make([][]string, len(records))
-	for i, record := range records {
-		expectedStrings[i] = getExpectedStringsForRecord(t, table, record)
-	}
-	return expectedStrings
-}
-
-func sortResults(table *schema.Table, records [][]string) {
-	cqIDIndex := table.Columns.Index(schema.CqIDColumn.Name)
+func sortResults(table *schema.Table, records []arrow.Record) {
+	idIndex := table.Columns.Index("id")
 	sort.Slice(records, func(i, j int) bool {
-		firstUUID := records[i][cqIDIndex]
-		secondUUID := records[j][cqIDIndex]
+		firstUUID := records[i].Column(idIndex).ValueStr(0)
+		secondUUID := records[j].Column(idIndex).ValueStr(0)
 		return strings.Compare(firstUUID, secondUUID) < 0
 	})
 }
@@ -176,17 +129,14 @@ func TestPlugin(t *testing.T) {
 		zerolog.ConsoleWriter{Out: os.Stderr, TimeFormat: time.StampMicro},
 	).Level(zerolog.DebugLevel).With().Timestamp().Logger()
 	p.SetLogger(l)
-	spec := specs.Source{
-		Name:         "test_mysql_source",
-		Path:         "cloudquery/mysql",
-		Version:      "vDevelopment",
-		Destinations: []string{"test"},
-		Tables:       []string{"test_mysql_source"},
-		Spec: client.Spec{
-			ConnectionString: getTestConnectionString(),
-		},
+	spec := client.Spec{
+		ConnectionString: getTestConnectionString(),
 	}
-	db, err := getTestDB(spec.Spec.(client.Spec).ConnectionString)
+	specBytes, err := json.Marshal(spec)
+	if err != nil {
+		t.Fatal(err)
+	}
+	db, err := getTestDB(spec.ConnectionString)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -200,8 +150,8 @@ func TestPlugin(t *testing.T) {
 		t.Fatal(err)
 	}
 	syncTime := time.Now()
-	data := schema.GenTestData(testTable, schema.GenTestDataOptions{MaxRows: 2, SyncTime: syncTime})
-	if err := insertTable(ctx, db, testTable, data); err != nil {
+	expectedRecords := schema.NewTestDataGenerator().Generate(testTable, schema.GenTestDataOptions{MaxRows: 2, SyncTime: syncTime})
+	if err := insertTable(ctx, db, testTable, expectedRecords); err != nil {
 		t.Fatal(err)
 	}
 
@@ -212,36 +162,66 @@ func TestPlugin(t *testing.T) {
 	if err := createTable(ctx, db, otherTable); err != nil {
 		t.Fatal(err)
 	}
-	otherData := schema.GenTestData(otherTable, schema.GenTestDataOptions{MaxRows: 1, SyncTime: syncTime})
+	otherData := schema.NewTestDataGenerator().Generate(otherTable, schema.GenTestDataOptions{MaxRows: 1, SyncTime: syncTime})
 	if err := insertTable(ctx, db, otherTable, otherData); err != nil {
 		t.Fatal(err)
 	}
 
 	// Init the plugin so we can call migrate
-	if err := p.Init(ctx, spec); err != nil {
+	if err := p.Init(ctx, specBytes, plugin.NewClientOptions{}); err != nil {
 		t.Fatal(err)
 	}
-	res := make(chan *schema.Resource, 1)
+	res := make(chan message.SyncMessage, 1)
 	g := errgroup.Group{}
 
 	g.Go(func() error {
 		defer close(res)
-		return p.Sync(ctx, syncTime, res)
+		opts := plugin.SyncOptions{Tables: []string{testTable.Name}, SkipTables: []string{otherTable.Name}}
+		return p.Sync(ctx, opts, res)
 	})
-	resources := make([]*schema.Resource, 0)
+	actualRecords := make([]arrow.Record, 0)
 	for r := range res {
-		resources = append(resources, r)
+		m, ok := r.(*message.SyncInsert)
+		if ok {
+			actualRecords = append(actualRecords, m.Record)
+		}
 	}
 	err = g.Wait()
 	if err != nil {
 		t.Fatal("got unexpected error:", err)
 	}
-	if len(resources) != 2 {
-		t.Fatalf("expected 2 resource, got %d", len(resources))
+	if len(actualRecords) != 2 {
+		t.Fatalf("expected 2 resource, got %d", len(actualRecords))
 	}
-	expectedStrings := getExpectedStrings(t, testTable, data)
-	actualStrings := getActualStrings(resources)
-	sortResults(testTable, actualStrings)
 
-	require.Equal(t, expectedStrings, actualStrings)
+	sortResults(testTable, actualRecords)
+
+	for recordIndex, expectedRecord := range expectedRecords {
+		actualRecord := actualRecords[recordIndex]
+		if expectedRecord.NumCols() != actualRecord.NumCols() {
+			t.Fatalf("expected record %d to have %d columns, got %d", recordIndex, expectedRecord.NumCols(), actualRecord.NumCols())
+		}
+		for columnIndex, col := range expectedRecord.Columns() {
+			actualColumn := actualRecord.Column(columnIndex)
+			columnName := expectedRecord.ColumnName(columnIndex)
+			if col.Len() != actualColumn.Len() {
+				t.Fatalf("expected record %d column %d (%s) to have length %d, got %d", recordIndex, columnIndex, columnName, col.Len(), actualColumn.Len())
+			}
+			for arrayIndex := 0; arrayIndex < col.Len(); arrayIndex++ {
+				var expectedValue, actualValue any
+				switch c := col.(type) {
+				case *types.JSONArray:
+					// TODO: Remove this when https://bugs.mysql.com/bug.php?id=98135 is fixed
+					// We can't do string comparison for JSON columns as MySQL saves them with extra spaces
+					expectedValue = c.Value(arrayIndex)
+					actualValue = actualColumn.(*types.JSONArray).Value(arrayIndex)
+				default:
+					expectedValue = col.ValueStr(arrayIndex)
+					actualValue = actualColumn.ValueStr(arrayIndex)
+				}
+
+				require.Equal(t, expectedValue, actualValue, "expected record %d column %d (%s) array index %d to have value %s, got %s", recordIndex, columnIndex, columnName, arrayIndex, expectedValue, actualValue)
+			}
+		}
+	}
 }