+++ plugins/source/snyk/client/testing.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cloudquery/plugin-sdk/v4/plugin"
 	"github.com/cloudquery/plugin-sdk/v4/scheduler"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/julienschmidt/httprouter"
@@ -43,15 +44,13 @@ func MockTestHelper(t *testing.T, table *schema.Table, createService func(*httpr
 		},
 	}
 
-	Transform(schema.Tables{table})
+	tables := schema.Tables{table}
+	Transform(tables)
 	table.IgnoreInTests = false
-	messages, err := sched.SyncAll(context.Background(), c, schema.Tables{table})
+
+	messages, err := sched.SyncAll(context.Background(), c, tables)
 	if err != nil {
 		t.Fatalf("failed to sync: %v", err)
 	}
-	records := messages.GetInserts().GetRecordsForTable(table)
-	emptyColumns := schema.FindEmptyColumns(table, records)
-	if len(emptyColumns) > 0 {
-		t.Fatalf("empty columns: %v", emptyColumns)
-	}
+	plugin.ValidateNoEmptyColumns(t, tables, messages)
 }